Add functionality for automatically cherry-picks to release branches
With this functionality the bot will parse the comments on a PR, and if the
comments match, and has an existing release-branch, then, if there is a match,
the bot will create a PR with the given cherry-pick (given no merge conflits).

On a conflict, the bot will simply show this in the result comment.

So input:

```
cherry pick to:
   * 2.4.x
   * 2.5.x
   * 2.6.x
```

output:

```
Hi :smileycat:
I did my very best, and this is the result of the cherry pick operation:
  * 2.4.x :white_check_mark: #PR-number
  * 2.5.x Had merge conflicts, you will have to fix this yourself :crying_cat_face:
  * 2.6.x :white_check_mark: #PR-number
```

The PR body of the opened cherries will have a mention of the comment author
upon creation.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>