gitlab_sync: Do not trigger CI build for mender-qa syncs
The build after sync for this repo (namely: on master) are virtually
always a waste of resources; any other repo (or in the worst case, in
the night) will be using mender-qa/master in the next integration
pipeline build.

And for the very few cases were it can be useful, is better to trigger
it manually.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>