cherry-pick: Notify if PR can be cherried upfront
This changes the comments when notifying reminding a user whether a cherry-pick
is needed from:

```
Hello smile_cat This PR contains changelog entries. Please, verify the need of
backporting it to the following release branches:
* 3.1.x (release 3.1.x)
* 3.0.x (release 3.0.x)
* 2.5.x (release 2.6.x)
```

To:

```
Hello smile_cat This PR contains changelog entries. Please, verify the need of
backporting it to the following release branches:
* 3.1.x (release 3.1.x) - :robot: :cherries:
* 3.0.x (release 3.0.x)
* 2.5.x (release 2.6.x) - :robot: :cherries:
```

In other words, adding the robot and cherry emojis to the branches which can be
automatically cherry-picked. In this way, the user is notified upfront whether
or not a cherry is possible without conflicts, and hence we do not need to try
cherry-picking to branches which we already know will cause conflicts.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>