menderqa_pipeline: Base the watch repositories list in release_tool list
So that we remove one point of duplication in our ecosystem.

The list was only being used to trigger in which repos to offer the
"start pipeline" functionality. Instead of a full list, hard-code only
the repos that are not part of the tool (namely: meta-mender).

The only drawback is that we are now suggesting in `gui` to start the
pipeline, but this might come in handy in the future if we integrate the
e2e tests with mender-qa.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>