fix: Do not process edited pull requests
Dependabot sometimes does a lot of edits in a short timespan. This causes our
test-bot to try and sync the branch on every request, basically doing a simple
DDos of our branch syncing to GitLab.

The fix is simple. Just ignore edits to PR's, which really is the right thing to
do anyways.

Changelog: None
Ticket: QA-538

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>