tests: Refresh the rest of the tests
Note that one has a minor adjustment due to the repo being out of the
release_tool but the change is irrelevant for the purpose or the test.

The others are just refreshing them with the automated output of the
Pytest plugin so that further runs don't show them in the diff.

Ticket: QA-794

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>