feat: Support for cryptographically sign Git commits
We opted for not modifying the `git` commands from the source code and
instead just externally configure Git to always sign commits.

With this approach we let the different deployments of this software to
opt-in the feature (by mounting the required secret) but let the
software run without it otherwise.

Ticket: SEC-1473

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>