fix: attempt to unprotect branches after deleting them
Just deleting the branch will still leave the protection rule in Gitlab.
When attempting to sync a branch while the protection is there, we will
get 409 conflict errors. Therefore we want to unprotect the branch after
deleting it to ensure we can re-create it and that the protection rule
is deleted after closing / merging a PR.

Changelog: None
Ticket: None

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>