chore: Replace source of truth for watch repos from release_tool to code
As per today, we have two opt-in pipelines that we can run through the
bot, the integration pipeline (iow from `integration` repository) or the
client pipeline (iow from the "client bundle" repositories).

Modify the code to reflect this:
* Disregard release_tool, use a hard-coded (but short) list of
  repositories instead.
* Rework the main comment to either offer the integration or the client
  pipeline as the main comment
* For suggesting cherry-picks, now we can only use the release tool on
  client repos. For others (mender-gateway at the time of writing) print
  a generic message instead.
* Note the TODOs added to the list of repos for further steps in the
  release split

I tried hard to add test for cherry-pick suggestions on non-client
repos but it has proved non feasible. On the unit tests we do run real
Git commands but we only access open sources repos and on acceptance
tests we run Git commands with dry run option.

Ticket: QA-1007

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>